Split browser provider implementations
